refactor: only show worflows and user's workflows + general code clean
